Consolidate shared types and rename numConstraints
Summary:
Move shared types (ConstraintData, ConstraintDataList, optional_ref,
VertexConstraintType enum and toString) from joint_error_function.h and
vertex_error_function.h into a new error_function_types.h header so they can
be used across all error function hierarchies without circular dependencies.
Rename numConstraints() to getNumConstraints() across all error function
classes for naming consistency with other getter methods.

Differential Revision: D94957203

+++ momentum/character_solver/distance_error_function.cpp
@@ -66,7 +66,7 @@ double DistanceErrorFunctionT<T>::getGradient(
     error += cons.weight * kDistanceWeight * this->weight_ * sqr(diff);
     const T wgt = T(2) * cons.weight * kDistanceWeight * this->weight_;
 
-    auto addGradient = [&](const int jFullBodyDOF, const Eigen::Vector3<T>& diff_p_world_cm) {
+    auto addGradient = [&](const size_t jFullBodyDOF, const Eigen::Vector3<T>& diff_p_world_cm) {
       // d(distance)/dTheta = d(distance)/d(p_world_cm) * d(p_world_cm)/dTheta
       // distance is defined as ||p_world_cm - cons.origin||, so
       // d(distance)/d(p_world_cm) =
